% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtgjsonsdk-package.R
\docType{package}
\name{mtgjsonsdk-package}
\alias{mtgjsonsdk}
\alias{mtgjsonsdk-package}
\title{mtgjsonsdk: DuckDB-Backed Query Client for MTGJSON Card Data}
\description{
Auto-downloads Parquet data from the MTGJSON CDN and exposes the full Magic: The Gathering dataset through R6-based query interfaces backed by DuckDB.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mtgjson.com}
  \item Report bugs at \url{https://github.com/mtgjson/mtgjson-sdk-r/issues}
}

}
\author{
\strong{Maintainer}: Zachary Halpern \email{zach@mtgjson.com}

Authors:
\itemize{
  \item Robert Pratt \email{rprat@users.noreply.github.com}
}

}
\keyword{internal}
